# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Auto Bot Meeting + Agent API
  version: 0.2.0
  description: >
    HTTP and WebSocket API for the voice-first meeting assistant and the
    Sprint 2.0 agent-first surface. Browser routes use HttpOnly session
    cookies (`auto_bot_session`); automation clients may use Bearer
    tokens when configured. The MCP tool surface that mirrors this API
    for external agents is documented in `docs/api/mcp-tools.md`; see
    `docs/adrs/0003-mcp-server-as-universal-external-surface.md` for the
    rationale.
servers:
  - url: http://localhost:3001
    description: Local Nova Sonic stack
  - url: http://localhost:3000
    description: Local OpenAI Realtime stack
security:
  - browserSession: []
  - bearerToken: []
paths:
...
  /tenant/pending_actions:
    get:
      summary: List the dry-run staging queue for the active tenant + board.
      description: >
        Returns `{actions: PendingAction[]}` newest-first. Terminal rows
        (applied / rejected / expired) are excluded by default; set
        `include_terminal=true` to surface the audit-log view. Source:
        `cmd/server/dry_run_http.go:17`-`:44` (`pendingActionsHandler`);
        store interface at `cmd/server/pending_actions.go:73`
        (`pendingActionStore.ListPendingActions`).
      parameters:
        - name: include_terminal
          in: query
          required: false
          schema: { type: boolean, default: false }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 50, maximum: 200 }
        - name: tenant_id
          in: query
          required: false
          schema: { type: string }
        - name: board_id
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Pending action list.
          content:
            application/json:
              schema:
                type: object
                required: [actions]
                properties:
                  actions:
                    type: array
                    items: { $ref: "#/components/schemas/PendingAction" }
        "401":
          description: Session or bearer token required.
        "503":
          description: Pending action store is not configured.
...
components:
  securitySchemes:
    browserSession:
      type: apiKey
      in: cookie
      name: auto_bot_session
    bearerToken:
      type: http
      scheme: bearer
    jiraWebhookSecret:
      type: http
      scheme: bearer
    jiraWebhookHeader:
      type: apiKey
      in: header
      name: X-Auto-Bot-Jira-Webhook-Secret
    slackSignature:
      type: apiKey
      in: header
      name: X-Slack-Signature
      description: >
        Slack-style HMAC of the request body keyed by
        `SLACK_SIGNING_SECRET`. The `X-Slack-Request-Timestamp` header is
        also required and must be within 5 minutes of the server clock.
        Verification at `internal/intake/slack.go`
        (`VerifySlackSignature`).
  schemas:
...
    PendingAction:
      type: object
      description: >
        A single staged tool call queued by dry-run mode. Mirrors
        `cmd/server/pending_actions.go:37` (`pendingAction`). Persisted
        in the `pending_actions` SQLite table; the in-memory store is
        used by tests.
      required: [tenant_id, board_id, action_id, tool, status]
      properties:
        tenant_id: { type: string }
        board_id:  { type: string }
        action_id: { type: string, description: "`pa_` + 24 hex characters." }
        tool:      { type: string, description: "cmd/server internal tool name (e.g. `create_ticket`, `move_ticket`)." }
        args:
          type: object
          additionalProperties: true
          description: Arbitrary JSON arguments; tool-specific shape.
        intent:
          type: object
          additionalProperties: true
          description: >
            Risk classification + evidence. The current implementation
            stores dispatcher / actor / call_id / transcript / raw_args
            (`cmd/server/dry_run.go:114`-`:120`). Schema will narrow as
            the risk classifier graduates.
        created_at: { type: string, description: RFC3339Nano. }
        expires_at: { type: string, description: RFC3339Nano deadline. Default TTL 24h. }
        status:
          type: string
          enum: [pending, applied, rejected, expired]
          description: Lifecycle state (`cmd/server/pending_actions.go:23`-`:27`).
        result:
          type: object
          additionalProperties: true
          description: Apply outcome (filled when `status=applied`).
        decision:
          $ref: "#/components/schemas/PendingActionDecision"
    PendingActionDecision:
      type: object
      description: >
        Resolution metadata for a `PendingAction`. Mirrors
        `cmd/server/pending_actions.go:54` (`pendingActionDecision`).
      required: [decided_at, disposition]
      properties:
        decided_at:  { type: string, description: RFC3339Nano. }
        decided_by:  { type: string, description: Operator identity (or `"sweeper"` for TTL expiry). }
        decided_via: { type: string, description: "Origin channel: `ui` | `voice` | `mcp` | `sweeper`." }
        disposition:
          type: string
          enum: [applied, rejected, expired]
        note: { type: string }
...